fix: disabled existing ood tests and added slidingsidebar test

fix: slidingsidebar test
